Fix disabling of incremental scala compilation
When disabling incremental compilation by setting force to true, the
analysis file was still created and loaded, causing Zinc to believe it
was running incrementally.
However the compilation output was effectively deleted. This resulted in
only producing output for changed types.
With the changes in this commit, the analysis information is ignored,
discarded if existing, and not preserved when incremental compilation is
disabled.

Fixes #13224